guides: command line guide and moving files
