+++ src/cli/generate/controller.js
