+++ src/cli/generate/model.js
