+++ src/cli/test.js
