--- src/core/server/index.js
@@ -1,7 +1,0 @@
-"use strict";
-
-module.exports = function() {
-  return {
-    start: require("./start")
-  };
-};
